refactor(#2055): decouple cache/ brick from nexus.core — zero import violations
Eliminate all 7 nexus.core import violations from the cache/ brick,
aligning it with the NEXUS LEGO Architecture (Tier 2 bricks must have
zero imports from kernel internals).

Phase 1: Move CacheStoreABC + NullCacheStore to cache/cache_store.py
  - Canonical home for the Fourth Pillar ABC (like Backend in backends/)
  - core/cache_store.py becomes backwards-compat re-export
  - Eliminates 4/7 violations (domain, inmemory, dragonfly, factory)

Phase 2: Move CachingBackendWrapper to backends/caching_wrapper.py
  - Backend decorator belongs with other backends (DelegatingBackend, etc.)
  - cache/backend_wrapper.py becomes backwards-compat re-export
  - Eliminates 2/7 violations (Backend, DelegatingBackend imports)

Phase 3: Move PostgresPersistentViewStore to storage/
  - Persistence concern with RecordStore affinity, not cache
  - cache/persistent_view_postgres.py becomes backwards-compat re-export
  - Eliminates 1/7 violation (persistent_view_store import)

Phase 4: Consolidate CacheBrick, deprecate CacheFactory
  - CacheBrick: backward-compat aliases, isinstance fix, async ctx mgr
  - CacheFactory: retained with deprecated docstring (Postgres fallback)
  - Remove redundant _startup_cache_factory from permissions.py
  - Fix compute_content_hash id(data) cache correctness bug

Phase 5: Add deadlock guard test for _l2_get_sync

All backwards-compat re-exports are identity-preserving (same object).
69 unit tests pass, 26 E2E server tests pass with permissions enabled.
