+++ src/nexus/cache/domain.py
@@ -26,7 +26,7 @@ import logging
 import struct
 from collections.abc import Awaitable, Callable
 
-from nexus.core.cache_store import CacheStoreABC
+from nexus.cache.cache_store import CacheStoreABC
 
 logger = logging.getLogger(__name__)
