+++ src/nexus/server/dependencies.py
@@ -22,7 +22,7 @@ from fastapi import Depends, Header, HTTPException, Request
 from nexus.server.token_utils import parse_sk_token
 
 if TYPE_CHECKING:
-    from nexus.core.cache_store import CacheStoreABC
+    from nexus.cache.cache_store import CacheStoreABC
 
 logger = logging.getLogger(__name__)
